Added Files from local computer
chore: Added all existing files from the local computer into the github repo.

This is done in order to get familiar with GitHub as well as upload all pre-existing files on the local machine as a form of starting with GitHub.  As of submission, the file "motion_planning.py" is used because of its relatively short time in order to find a path from a starting position to a goal position.  However, the file "motion_planning_prob.py" file tackles the same problem using the probabilistic roadmap approach.