Created README
docs: Transferring the documentation written as a PDF format to a markdown format. 

As of now, all of the detailed explanation has not been translated into a markdown yet.  Firstly, the PDF format is different and also a better presentation style needs to be implemented and hence needs more commits following this one.